Refactor out run functions
